feat: make  public
