+++ torchopt/diff/implicit/decorator.py
@@ -36,7 +36,7 @@ from torchopt.typing import (
 )
 
 
-__all__ = ['custom_root', 'root_vjp']
+__all__ = ['custom_root']
 
 
 Args = Tuple[Any, ...]
@@ -416,55 +416,6 @@ def _custom_root(
         return output
 
     return wrapped_solver_fn
-
-
-# pylint: disable-next=too-many-arguments,too-many-locals,too-many-branches
-def root_vjp(
-    optimality_fn: Callable[..., TensorOrTensors],
-    solution: TupleOfTensors,
-    args: Args,
-    grad_outputs: TupleOfTensors,
-    output_is_tensor: bool,
-    argnums: tuple[int, ...],
-    solve: Callable[..., TensorOrTensors] = None,
-) -> TupleOfOptionalTensors:
-    """Return vector-Jacobian product of a root.
-
-    The root is the `solution` of ``optimality_fn(solution, *args) == 0``.
-
-    Args:
-        optimality_fun (callable): An equation function, ``optimality_fn(params, *args)``. The
-            invariant is ``optimality_fn(solution, *args) == 0`` at ``solution``.
-        solution (tuple of Tensors): solution / root of `optimality_fun`.
-        args (Args): tuple containing the arguments with respect to which we wish to
-            differentiate ``solution`` against.
-        grad_outputs (tuple of Tensors): The "vector" in the vector-Jacobian product.
-            Usually gradients w.r.t. each output. None values can be specified for scalar
-            Tensors or ones that don't require grad. If a None value would be acceptable
-            for all grad_tensors, then this argument is optional. Default: None.
-        output_is_tensor (bool): Whether the output of ``optimality_fn`` is a single tensor.
-        argnums (int or tuple of int): Specifies arguments to compute gradients with respect to. The
-            ``argnums`` can be an integer or a tuple of integers.
-        solve (callable, optional): A linear solver of the form ``solve(matvec, b)``.
-            (default: :func:`linear_solve.solve_normal_cg`)
-
-    Returns:
-        tuple of the same length as ``len(args)`` containing the vector-Jacobian products w.r.t.
-        each argument. Each ``vjps[i]`` has the same pytree structure as
-        ``args[i]``.
-    """
-    if solve is None:
-        solve = linear_solve.solve_normal_cg()
-
-    return _root_vjp(
-        optimality_fn=optimality_fn,
-        solution=solution,
-        args=args,
-        grad_outputs=grad_outputs,
-        output_is_tensor=output_is_tensor,
-        argnums=argnums,
-        solve=solve,
-    )
 
 
 def custom_root(
